refactor: vacancies CRUD methods
