+++ src/api/Vacancies/activeVacancy.js
@@ -1,0 +1,5 @@
+import axiosClient from '../axiosClient';
+
+export const activeVacancy = (id) => {
+    return axiosClient.patch(`/vacancies-status-active/${id}`);
+};
