+++ src/api/Vacancies/allVacancies.js
@@ -1,9 +1,5 @@
-import axios from 'axios';
+import axiosClient from '../axiosClient';
 
-export const getJobs = async () => {
-    const {
-        data: { data },
-    } = await axios.get('https://gethiredplatzi.herokuapp.com/api/v1/vacancies');
-    const statusActive = data?.filter((vacancy) => vacancy.status === true);
-    return statusActive;
+export const getAllVacancy = () => {
+    return axiosClient.get('/vacancies');
 };
