+++ src/api/Vacancies/inactiveVacancy.js
@@ -1,0 +1,5 @@
+import axiosClient from '../axiosClient';
+
+export const inactiveVacancy = (id) => {
+    return axiosClient.patch(`/vacancies-status-inactive/${id}`);
+};
